fix: Update API endpoints in code changes to match new API structure
