+++ components/features/profileCompletion/addServices.tsx
@@ -3,11 +3,8 @@ import { useSelector, useDispatch } from "react-redux";
 import { AiOutlineClose } from "react-icons/ai";
 import AddServicesForm from "../../forms/addServicesForm";
 import ProfileProgress from "@/components/core/cards/progress";
-import {
-  useAssignService,
-  useGetAllServices,
-  useGetServiceCategories,
-} from "@/app/api/requests";
+import { useAssignService, useGetAllServices } from "@/app/api/businesses";
+import { useGetServiceCategories } from "@/app/api/services";
 import { RootState } from "@/store/store";
 import { setQueuedServices, setStep } from "@/store/completeProfileSlice";
 import Toast from "@/components/shared/toasts/authToast";
